requirements.txt
asgiref==3.2.7
Django==2.2.13
djangorestframework==3.11.0
mysqlclient==1.4.6
pytz==2020.1
sqlparse==0.3.1
